+++ examples/objdump.rs
@@ -1,13 +1,18 @@
-extern crate macho;
 extern crate capstone;
+extern crate macho;
+
 use std::env;
 use std::fs;
 use std::io::Read;
 use std::process;
+use capstone::prelude::*;
 
 fn main() {
-    let cs = capstone::Capstone::new(capstone::Arch::X86,
-                                     capstone::Mode::Mode64).ok().unwrap();
+    let cs = Capstone::new()
+        .x86()
+        .mode(arch::x86::ArchMode::Mode64)
+        .build()
+        .expect("Failed to create capstone handle");
 
     let args: Vec<_> = env::args().collect();
     if args.len() != 2 {
@@ -25,7 +30,8 @@ fn main() {
         if segment.segname == "__TEXT" {
             for section in segment.sections {
                 if section.sectname == "__text" {
-                    let text = &buf[section.offset as usize .. (section.offset as u64 + section.size) as usize];
+                    let text = &buf[section.offset as usize..
+                                        (u64::from(section.offset) + section.size) as usize];
                     match cs.disasm_all(text, section.addr) {
                         Ok(insns) => {
                             println!("Got {} instructions", insns.len());
@@ -33,7 +39,7 @@ fn main() {
                             for i in insns.iter() {
                                 println!("{}", i);
                             }
-                        },
+                        }
                         Err(err) => {
                             println!("Error: {}", err);
                             process::exit(1);