+++ src/constants.rs
@@ -1,6 +1,7 @@
 use capstone_sys::*;
 use capstone_sys::cs_arch::*;
 use capstone_sys::cs_mode::*;
+use capstone_sys::cs_opt_value::*;
 use std::convert::From;
 
 /// Define an `enum` that corresponds to a capstone enum
@@ -9,10 +10,7 @@ use std::convert::From;
 macro_rules! define_cs_enum_wrapper {
     ( [
         $( #[$enum_attr:meta] )*
-        => $rust_enum:ident = $cs_enum:ident,
-        $( #[$cs_to_rust_attrs:meta] )*
-        ,
-        $( #[$rust_to_cs_attrs:meta] )*
+        => $rust_enum:ident = $cs_enum:ident
       ]
       $( $( #[$attr:meta] )*
       => $rust_variant:ident = $cs_variant:tt; )* ) => {
@@ -26,19 +24,7 @@ macro_rules! define_cs_enum_wrapper {
             )*
         }
 
-        $( #[$cs_to_rust_attrs] )*
-        impl From<$cs_enum> for $rust_enum {
-            fn from(other: $cs_enum) -> Self {
-                match other {
-                    $(
-                        $cs_enum::$cs_variant => $rust_enum::$rust_variant,
-                    )*
-                }
-            }
-        }
-
-        $( #[$rust_to_cs_attrs] )*
-        impl From<$rust_enum> for $cs_enum {
+        impl ::std::convert::From<$rust_enum> for $cs_enum {
             fn from(other: $rust_enum) -> Self {
                 match other {
                     $(
@@ -53,8 +39,7 @@ macro_rules! define_cs_enum_wrapper {
 define_cs_enum_wrapper!(
     [
         /// Architectures for the disassembler
-        => Arch = cs_arch,
-        ,
+        => Arch = cs_arch
     ]
     /// ARM (Advanced RISC Machine)
     => ARM = CS_ARCH_ARM;
@@ -72,22 +57,15 @@ define_cs_enum_wrapper!(
     => SYSZ = CS_ARCH_SYSZ;
     /// XCore
     => XCORE = CS_ARCH_XCORE;
-    => MAX = CS_ARCH_MAX;
-    /// used for `cs_support()`
-    => ALL = CS_ARCH_ALL;
 );
 
 define_cs_enum_wrapper!(
     [
         /// Disassembler modes
-        => Mode = cs_mode,
-        #[cfg(other)]
-        ,
+        => Mode = cs_mode
     ]
-    /// little-endian mode
-    => LittleEndian = CS_MODE_LITTLE_ENDIAN;
     /// 32-bit ARM
-    => Arm32 = CS_MODE_ARM;
+    => Arm = CS_MODE_ARM;
     /// 16-bit mode (X86)
     => Mode16 = CS_MODE_16;
     /// 32-bit mode (X86)
@@ -96,12 +74,6 @@ define_cs_enum_wrapper!(
     => Mode64 = CS_MODE_64;
     /// ARM's Thumb mode, including Thumb-2
     => Thumb = CS_MODE_THUMB;
-    /// ARM's Cortex-M series
-    => MClass = CS_MODE_MCLASS;
-    /// ARMv8 A32 encodings for ARM
-    => V8 = CS_MODE_V8;
-    /// MicroMips mode (MIPS)
-    => Micro = CS_MODE_MICRO;
     /// Mips III ISA
     => Mips3 = CS_MODE_MIPS3;
     /// Mips32r6 ISA
@@ -110,12 +82,45 @@ define_cs_enum_wrapper!(
     => MipsGP64 = CS_MODE_MIPSGP64;
     /// SparcV9 mode (Sparc)
     => V9 = CS_MODE_V9;
-    /// big-endian mode
-    => BigEndian = CS_MODE_BIG_ENDIAN;
-    /// Mips32 ISA (Mips)
-    => Mips32 = CS_MODE_MIPS32;
-    /// Mips64 ISA (Mips)
-    => Mips64 = CS_MODE_MIPS64;
+    /// Default mode for little-endian
+    => Default = CS_MODE_LITTLE_ENDIAN;
+);
+
+define_cs_enum_wrapper!(
+    [
+        /// Extra modes or features that can be enabled with some modes
+        => ExtraMode = cs_mode
+    ]
+    /// ARM's Cortex-M series. Works with `Arm` mode.
+    => MClass = CS_MODE_MCLASS;
+    /// ARMv8 A32 encodings for ARM. Works with `Arm` and `Thumb` modes.
+    => V8 = CS_MODE_V8;
+    /// MicroMips mode. Works in `MIPS` mode.
+    => Micro = CS_MODE_MICRO;
+);
+
+define_cs_enum_wrapper!(
+    [
+        /// Disassembler endianness
+        => Endian = cs_mode
+    ]
+    /// Little-endian mode
+    => Little = CS_MODE_LITTLE_ENDIAN;
+    /// Big-endian mode
+    => Big = CS_MODE_BIG_ENDIAN;
+);
+
+define_cs_enum_wrapper!(
+    [
+        /// Disassembly syntax
+        => Syntax = cs_opt_value
+    ]
+    /// Intel syntax
+    => Intel = CS_OPT_SYNTAX_INTEL;
+    /// AT&T syntax (also known as GNU assembler/GAS syntax)
+    => Att = CS_OPT_SYNTAX_ATT;
+    /// No register name
+    => NoRegName = CS_OPT_SYNTAX_NOREGNAME;
 );
 
 pub(crate) struct OptValue(pub cs_opt_value);
@@ -127,5 +132,21 @@ impl From<bool> for OptValue {
         } else {
             OptValue(cs_opt_value::CS_OPT_OFF)
         }
+    }
+}
+
+/// Representation of `cs_mode`. We use this to have a type that we can transmute() to that has
+/// the same memory representation.
+pub(crate) type CsModeRepr = i32;
+
+#[cfg(test)]
+mod test {
+    use capstone_sys::cs_mode;
+    use std::mem;
+    use super::CsModeRepr;
+
+    #[test]
+    fn test_cs_mode_size() {
+        assert_eq!(mem::size_of::<cs_mode>(), mem::size_of::<CsModeRepr>());
     }
 }